+++ CHANGELOG.md
@@ -6,12 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
-## [2.0.1] - 2026-09-22
+## [2.0.1]
 
 ### Major Fixes
 - Fixed a bug that led to showing the activity screen while navigating rather than novel screen
 - Downloaded chapters can now be backed and restored properly
 - Repository URL can now be removed at the onboarding for a completely blank experience
+- Storage location selected in the new app does not change even if old backup was brought back
 
 ### Added
 - Brought back the cancellation box to confirm before cancelling or deleting downloaded chapters